+++ src/Controller/Admin/ChartsController.php
@@ -8,7 +8,9 @@ use App\Block\Last12MonthsTopTenCustomerEarningsBlock;
 use App\Block\LastMonthsInvoicingResumeBlock;
 use App\Block\TopTenCustomerEarningsBlock;
 use App\Block\TotalRetainingEarningsBlock;
+use App\Manager\GoogleAnalyticsManager;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
@@ -65,5 +67,26 @@ final class ChartsController extends AbstractController
                 'bottom' => [],
             ],
         ]);
+    }
+
+    #[Route('/google-user-oauth-token', name: 'admin_google_user_oauth_token')]
+    public function googleUserOauthToken(Request $request, GoogleAnalyticsManager $gam): Response
+    {
+        $authCode = $request->get('code');
+        if ($authCode) {
+            $accessToken = $gam->getGoogleApiClient()->fetchAccessTokenWithAuthCode($authCode);
+            // check to see if there was an error
+            if (array_key_exists('error', $accessToken)) {
+                throw $this->createNotFoundException(implode(', ', $accessToken));
+            }
+            $gam->getGoogleApiClient()->setAccessToken($accessToken);
+            $gam->setGoogleAnalyticsServiceByGoogleApiClient($gam->getGoogleApiClient());
+            $gam->setGoogleUserAccessToken($accessToken);
+        } else {
+            throw $this->createAccessDeniedException('No auth code URL param found');
+        }
+
+        return $this->redirectToRoute('sonata_admin_dashboard');
+
     }
 }