+++ src/Manager/AssetsManager.php
@@ -12,12 +12,19 @@ class AssetsManager
     public const MIME_APPLICATION_PDF_X_TYPE = 'application/x-pdf';
 
     private Filesystem $filesystem;
+    private string $projectRootPathDir;
     private string $publicPathDir;
 
     public function __construct(string $kpd)
     {
         $this->filesystem = new Filesystem();
-        $this->publicPathDir = $kpd.DIRECTORY_SEPARATOR.'public';
+        $this->projectRootPathDir = $kpd.DIRECTORY_SEPARATOR;
+        $this->publicPathDir = $this->projectRootPathDir.'public';
+    }
+
+    public function getProjectRootDir(): string
+    {
+        return $this->projectRootPathDir;
     }
 
     public function getLocalPublicPath(string $path): string
@@ -28,5 +35,10 @@ class AssetsManager
         }
 
         return $publicPath;
+    }
+
+    public function fileExists(string $filepath): bool
+    {
+        return $this->filesystem->exists($filepath);
     }
 }