+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog
  * keep working
  * add total projects dashboard widget
  * add total contact messages dashboard widget
+ * add Google Analytics API bridge
 
 ##### Version 1.02.001 (2023-0618)
  * add Enterprise management